+++ packages/auth-client/package.json
@@ -48,7 +48,6 @@
   },
   "peerDependencies": {
     "@dfinity/agent": "^0.14.1",
-    "@dfinity/authentication": "^0.14.1",
     "@dfinity/identity": "^0.14.1",
     "@dfinity/principal": "^0.14.1"
   },
