--- packages/authentication/jest.config.ts
@@ -1,18 +1,0 @@
-import baseConfig from '../../jest.config.base';
-const packageName = 'authentication';
-
-module.exports = {
-  ...baseConfig,
-  roots: [`<rootDir>/packages/${packageName}`],
-  bail: false,
-  moduleDirectories: ['node_modules'],
-  modulePaths: [`<rootDir>/packages/${packageName}/src/`],
-  setupFiles: [`<rootDir>/packages/${packageName}/test-setup.ts`],
-  transform: {
-    '^.+\\.ts$': 'ts-jest',
-  },
-  collectCoverageFrom: ['src/**/*.{ts,tsx}'],
-  name: packageName,
-  displayName: packageName,
-  rootDir: '../..',
-};
